Update: ValueObjects implementados en Atriculo

+++ LogicaAplicacion/DataTransferObjects/Models/Articulos/ListadoArticulosDTO.cs
@@ -12,7 +12,7 @@ namespace LogicaAplicacion.DataTransferObjects.Models.Articulos
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
         public long Codigo { get; set; }
-        public int PrecioVenta { get; set; }
+        public decimal PrecioVenta { get; set; }
         public int Stock { get; set; }
 
         public ArticulosListadoDTO() { }
